requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
mysqlclient==2.2.4
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.1